\name{bincombinations}
\title{Binary Combinations}
\usage{
bincombinations(p)
}
\alias{bincombinations}
\arguments{
 \item{p}{Length of binary vectors}
}
\description{
  Returns a matrix containing the \eqn{2^p} vectors of length \code{p}.
}
\author{Friedrich Leisch}
\examples{
bincombinations(2)
bincombinations(3)
}
\keyword{utilities}
