\name{e1071-internal}
\alias{prune.bclust}
\alias{scale.data.frame}
\alias{na.fail.matrix.csr}
\title{Internal e1071 functions}
\description{
  Internal e1071 functions.
}
\usage{
prune.bclust(object, x, minsize=1, dohclust=FALSE, ...)}
scale.data.frame(x, scale = TRUE, center = TRUE)
na.fail.matrix.csr(object, ...)
\details{
  These are not to be called by the user.
}
\keyword{internal}
