\name{HouseVotes84}
\alias{HouseVotes84}
\title{United States Congressional Voting Records 1984}
\usage{data(HouseVotes84)}
\description{
    This data set includes votes for each of the U.S. House of
    Representatives Congressmen on the 16 key votes identified by the
    CQA.  The CQA lists nine different types of votes: voted for, paired
    for, and announced for (these three simplified to yea), voted
    against, paired against, and announced against (these three
    simplified to nay), voted present, voted present to avoid conflict
    of interest, and did not vote or otherwise make a position known
    (these three simplified to an unknown disposition).
}
\keyword{datasets}
\format{
    A data frame with 435 observations on 17 variables:
    \tabular{rl}{
   1 \tab Class Name: 2 (democrat, republican)\cr
   2 \tab handicapped-infants: 2 (y,n)\cr
   3 \tab water-project-cost-sharing: 2 (y,n)\cr
   4 \tab adoption-of-the-budget-resolution: 2 (y,n)\cr
   5 \tab physician-fee-freeze: 2 (y,n)\cr
   6 \tab el-salvador-aid: 2 (y,n)\cr
   7 \tab religious-groups-in-schools: 2 (y,n)\cr
   8 \tab anti-satellite-test-ban: 2 (y,n)\cr
   9 \tab aid-to-nicaraguan-contras: 2 (y,n)\cr
  10 \tab mx-missile: 2 (y,n)\cr
  11 \tab immigration: 2 (y,n)\cr
  12 \tab synfuels-corporation-cutback: 2 (y,n)\cr
  13 \tab education-spending: 2 (y,n)\cr
  14 \tab superfund-right-to-sue: 2 (y,n)\cr
  15 \tab crime: 2 (y,n)\cr
  16 \tab duty-free-exports: 2 (y,n)\cr
  17 \tab export-administration-act-south-africa: 2 (y,n)\cr
  }
}
\source{
    \itemize{
	\item Source: Congressional Quarterly Almanac, 98th Congress,
	2nd session 1984, Volume XL: Congressional Quarterly Inc.,
	ington, D.C., 1985
	\item Donor: Jeff Schlimmer (Jeffrey.Schlimmer@a.gp.cs.cmu.edu)
    }

    These data have been taken from the UCI Repository Of Machine Learning
    Databases at
    \itemize{
      \item \url{ftp://ftp.ics.uci.edu/pub/machine-learning-databases}
      \item \url{http://www.ics.uci.edu/~mlearn/MLRepository.html}
    }
    and were converted to R format by Friedrich.Leisch@ci.tuwien.ac.at.
}
    
