\name{Ozone}
\alias{Ozone}
\title{Los Angeles Ozone Pollution Data, 1976}
\usage{data(Ozone)}
\keyword{datasets}
\description{A data frame with 366 observations on 13 variables, each
  observation is one day}
\format{
    \tabular{rl}{
   1 \tab Month: 1 = January, ..., 12 = December\cr
   2 \tab Day of month\cr
   3 \tab Day of week: 1 = Monday, ..., 7 = Sunday\cr
   4 \tab Daily maximum one-hour-average ozone reading\cr
   5 \tab 500 millibar pressure height (m) measured at Vandenberg AFB\cr
   6 \tab Wind speed (mph) at Los Angeles International Airport (LAX)\cr
   7 \tab Humidity (\%) at LAX\cr
   8 \tab Temperature (degrees F) measured at Sandburg, CA\cr
   9 \tab Temperature (degrees F) measured at El Monte, CA\cr
  10 \tab Inversion base height (feet) at LAX\cr
  11 \tab Pressure gradient (mm Hg) from LAX to Daggett, CA\cr
  12 \tab Inversion base temperature (degrees F) at LAX\cr
  13 \tab Visibility (miles) measured at LAX\cr
  }
}
\details{
The problem is to predict the daily maximum one-hour-average
ozone reading (V4).
}
\source{
    Leo Breiman, Department of Statistics, UC Berkeley.  Data used in
    Leo Breiman and Jerome H. Friedman (1985), Estimating optimal
    transformations for multiple regression and correlation, JASA, 80, pp.
    580-598.
}
    
