\name{permutations}
\alias{permutations}
\title{All Permutations of Integers 1:n}
\description{
  Returns a matrix containing all permutations of the integers
  \code{1:n} (one permutation per row).
}
\usage{
permutations(n)
}
\arguments{
  \item{n}{Number of element to permute.}
}
\author{Friedrich Leisch}
\examples{
permutations(3)
}
\keyword{datagen}
