\name{e1071-deprecated}
\alias{e1071-deprecated}
\title{Deprecated Functions in Package e1071}
\description{
  These functions are provided for compatibility with older versions of
  package \pkg{e1071} only, and may be defunct as soon as of the next
  release.
}
%\usage{
%}
\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
